Create frontend-O-matic.ahk
This is just a quick and dirty basic file creation tool. 
Its purpose is to display whatever on the fly as needed. 
Add a server file to make it a file server for more fun or use as is.
You will have to go in and manually edit the HTML/CSS/JS individual files. Add your own stuff into the storage as well.